Response is not accepted for any key press

Response is not accepted for any key press

respond_no_key()

Details

Many of the functions within the trial_ family are designed to allow participants to respond using a key press, generally by specifying a a choices argument that indicates which keys will be accepted as valid responses (e.g., choices = c("f","j")). There are also cases where you may wish to disable this, so that no key presses will be counted as valid responses (e.g., when a trial runs for a fixed duration but no response is expected). In those cases, specifying choices = respond_no_key() will produce the desired behaviour.

See also