Return a javascript function to save data via a script on the webserver

Return a javascript function to save data via a script on the webserver

save_webserver()

Value

Return a javascript function to save data via a script on the webserver

Details

The purpose of the save_webserver() is to return a javascript function that, when called from within the jsPsych experiment, will write the data to the server. This assumes that the experiment will be run on a (php)script-enabled webserver. This way, you know the data will never touch any other computer than the server you've presumably secured and have data processing agreements in place for.

See also