Displays a sequence of images at a fixed rate

Displays a sequence of images at a fixed rate

trial_animation(stimuli, frame_time = 250, frame_isi = 0,
  sequence_reps = 1, choices = respond_any_key(), prompt = NULL,
  post_trial_gap = 0, on_finish = NULL, on_load = NULL,
  data = NULL)

Arguments

stimuli
A vector of paths to the image files

frame_time
How long to display each image, in milliseconds

frame_isi
How long is the gap between images, in milliseconds

sequence_reps
How many times to repeat the sequence

choices
A character vector of keycodes (either numeric values or the characters themselves). Alternatively, respond_any_key() and respond_no_key() can be used

prompt
A string (may contain HTML) that will be displayed below the stimulus, intended as a reminder about the actions to take (e.g., which key to press).

post_trial_gap
The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.

on_finish
A javascript callback function to execute when the trial finishes

on_load
A javascript callback function to execute when the trial begins, before any loading has occurred

data
An object containing additional data to store for the trial

Value

Functions with a trial_ prefix always return a "trial" object. A trial object is simply a list containing the input arguments, with NULL elements removed. Logical values in the input (TRUE and FALSE) are transformed to character vectors "true" and "false" and are specified to be objects of class "json", ensuring that they will be written to file as the javascript logicals, true and false.

Details

In addition to the default data collected by all plugins, this plugin collects the following data for each trial.

The animation_sequence value is an array encoded in JSON format. Each element of the array is an object that represents a stimulus in the animation sequence. Each object has a stimulus property, which is the image that was displayed, and a time property, which is the time in ms, measured from when the sequence began, that the stimulus was displayed.

The responses value is an array encoded in JSON format. Each element of the array is an object representing a response given by the subject. Each object has a stimulus property, indicating which image was displayed when the key was pressed, an rt property, indicating the time of the key press relative to the start of the animation, and a key_press property, indicating which key was pressed.