Displays an image, records responses with the keyboard and gives feedback

Displays an image, records responses with the keyboard and gives feedback

trial_categorize_image(stimulus, key_answer, choices = respond_any_key(),
  text_answer = "", correct_text = "Correct",
  incorrect_text = "Wrong", prompt = NULL,
  force_correct_button_press = FALSE, show_stim_with_feedback = TRUE,
  show_feedback_on_timeout = FALSE,
  timeout_message = "Please respond faster", stimulus_duration = NULL,
  feedback_duration = 2000, trial_duration = NULL,
  post_trial_gap = 0, on_finish = NULL, on_load = NULL,
  data = NULL)

Arguments

stimulus
The path to the image file to be displayed.

key_answer
The numeric key code indicating the correct response

choices
A character vector of keycodes (either numeric values or the characters themselves). Alternatively, respond_any_key() and respond_no_key() can be used

text_answer
A label associated with the correct answer

correct_text
Text to display when correct answer given ('%ANS%' substitutes text_answer)

incorrect_text
Text to display when wrong answer given ('%ANS%' substitutes text_answer)

prompt
A string (may contain HTML) that will be displayed below the stimulus, intended as a reminder about the actions to take (e.g., which key to press).

force_correct_button_press
If TRUE the correct button must be pressed after feedback in order to advance

show_stim_with_feedback
If TRUE the stimulus image will be displayed as part of the feedback. Otherwise only text is shown

show_feedback_on_timeout
If TRUE the "wrong answer" feedback will be presented on timeout. If FALSE, a timeout message is shown

timeout_message
The message to show on a timeout non-response

stimulus_duration
How long to show the stimulus, in milliseconds. If NULL, then the stimulus will be shown until the subject makes a response

feedback_duration
How long to show the feedback, in milliseconds

trial_duration
How long to wait for a response before ending trial in milliseconds. If NULL, the trial will wait indefinitely. If no response is made before the deadline is reached, the response will be recorded as NULL.

post_trial_gap
The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.

on_finish
A javascript callback function to execute when the trial finishes

on_load
A javascript callback function to execute when the trial begins, before any loading has occurred

data
An object containing additional data to store for the trial

Details

In addition to the default data collected by all plugins, this plugin collects the following data for each trial:

The stimulus value is the path to the image file. The key_press value indicates which key the subject pressed. The value is the numeric key code corresponding to the subject's response. The rt value is the response time in milliseconds for the subject to make a response. The time is measured from when the stimulus first appears on the screen until the subject's response. The correct value is true if the subject got the correct answer, false otherwise.