Displays an image stimulus and records responses generated with a button click

Displays an image stimulus and records responses generated with a button click

trial_image_button_response(stimulus, stimulus_height = NULL,
  stimulus_width = NULL, maintain_aspect_ratio = TRUE,
  choices = c("button 0", "button 1", "button 2"), button_html = NULL,
  margin_vertical = "0px", margin_horizontal = "8px", prompt = NULL,
  stimulus_duration = NULL, trial_duration = NULL,
  response_ends_trial = TRUE, post_trial_gap = 0, on_finish = NULL,
  on_load = NULL, data = NULL)

Arguments

stimulus
The path of the image file to be displayed.

stimulus_height
Set the height of the image in pixels. If NULL, then the image will display at its natural height.

stimulus_width
Set the width of the image in pixels. If NULL, then the image will display at its natural width.

maintain_aspect_ratio
If setting only the width or only the height and this parameter is TRUE, then the other dimension will be scaled to maintain the image's aspect ratio.

choices
Labels for the buttons. Each element of the character vector will generate a different button.

button_html
A template of HTML for generating the button elements (see details)

margin_vertical
Vertical margin of the buttons

margin_horizontal
Horizontal margin of the buttons

prompt
A string (may contain HTML) that will be displayed below the stimulus, intended as a reminder about the actions to take (e.g., which key to press).

stimulus_duration
How long to show the stimulus, in milliseconds. If NULL, then the stimulus will be shown until the subject makes a response

trial_duration
How long to wait for a response before ending trial in milliseconds. If NULL, the trial will wait indefinitely. If no response is made before the deadline is reached, the response will be recorded as NULL.

response_ends_trial
If TRUE, then the trial will end when a response is made (or the trial_duration expires). If FALSE, the trial continues until the deadline expires.

post_trial_gap
The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.

on_finish
A javascript callback function to execute when the trial finishes

on_load
A javascript callback function to execute when the trial begins, before any loading has occurred

data
An object containing additional data to store for the trial

Details

This plugin displays HTML content and records responses generated by button click. The stimulus can be displayed until a response is given, or for a pre-determined amount of time. The trial can be ended automatically if the subject has failed to respond within a fixed length of time.

The response buttons can be customized using HTML formatting, via the button_html argument. This argument allows the user to specify an HTML used to generating the button elements. If this argument is a vector of the same length as choices then the i-th element of button_html will be used to define the i-th response button. If button_html is a single string then the same template will be applied to every button. The templating is defined uing a special string " the jsPsych library creates an HTML button of class "jspsych-btn" and the styling is governed by the corresponding CSS.

In addition to the default data collected by all plugins, this plugin collects the following data for each trial: the rt value is the response time in milliseconds taken for the user to make a response. The time is measured from when the stimulus first appears on the screen until the response; button_pressed is a numeric value indicating which button was pressed. The first button in the choices array is recorded as value 0, the second is value 1, and so on; stimulus The HTML content that was displayed on the screen.