Specify a video trial with keyboard response

The trial_video_keyboard_response function is used to play a video stimulus and collect a response using a key press.

trial_video_keyboard_response(sources, trial_ends_after_video = FALSE,
  width = NULL, height = NULL, autoplay = TRUE, controls = FALSE,
  start = NULL, stop = NULL, rate = 1, choices = respond_any_key(),
  prompt = NULL, trial_duration = NULL, response_ends_trial = TRUE,
  post_trial_gap = 0, on_finish = NULL, on_load = NULL,
  data = NULL)

Arguments

sources
Path(s) to the video file. Videos may be specified in multiple formats (e.g., .mp4, .ogg, .webm)

trial_ends_after_video
If TRUE the trial will end as soon as the video finishes playing.

width
The width of the video display in pixels (if NULL, natural width is used)

height
The height of the video display in pixels (if NULL, natural height is used)

autoplay
Does the video play automatically?

controls
Should the video controls be made available to the user?

start
Specifies the time in seconds to start the video (if NULL, start at beginning)

stop
Specifies the time in seconds to stop the video (if NULL, stop at end)

rate
What rate to play the video (1 = normal, <1 slower, >1 faster)

choices
A character vector of keycodes (either numeric values or the characters themselves). Alternatively, respond_any_key() and respond_no_key() can be used

prompt
A string (may contain HTML) that will be displayed below the stimulus, intended as a reminder about the actions to take (e.g., which key to press).

trial_duration
How long to wait for a response before ending trial in milliseconds. If NULL, the trial will wait indefinitely. If no response is made before the deadline is reached, the response will be recorded as NULL.

response_ends_trial
If TRUE, then the trial will end when a response is made (or the trial_duration expires). If FALSE, the trial continues until the deadline expires.

post_trial_gap
The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.

on_finish
A javascript callback function to execute when the trial finishes

on_load
A javascript callback function to execute when the trial begins, before any loading has occurred

data
An object containing additional data to store for the trial

Value

Functions with a trial_ prefix always return a "trial" object. A trial object is simply a list containing the input arguments, with NULL elements removed. Logical values in the input (TRUE and FALSE) are transformed to character vectors "true" and "false" and are specified to be objects of class "json", ensuring that they will be written to file as the javascript logicals, true and false.

Details

This plugin displays an HTML stimulus and responses generated by a key press. The stimulus can be displayed until a response is given, or for a pre-determined amount of time. The trial can be ended automatically if the subject has failed to respond within a fixed length of time.

The data recorded by this trial is as follows:

The rt value is the response time in milliseconds taken for the user to make a response. The time is measured from when the stimulus first appears on the screen until the response.

The key_press variable is the numeric javascript key code corresponding to the response.

The stimulus variable records a JSON encoding of the sources array.

In addition, it records default variables that are recorded by all trials:

trial_type is a string that records the name of the plugin used to run the trial.

trial_index is a number that records the index of the current trial across the whole experiment.

time_elapsed counts the number of milliseconds since the start of the experiment when the trial ended.

internal_node_id is a string identifier for the current "node" in the timeline.